feat(suite): route homescreen change through connect-flow useConnect
Bypasses the redux modal context for the device-background-gallery flow:
- HomescreenGallery now drives applySettings via useConnect, rendering
  ConnectSubprocessModal (PIN/passphrase/confirmation/no-backup) locally
- ChangeHomescreenButtons opens BackgroundGalleryModal via local state
- Removes the 'device-background-gallery' UserContextPayload variant and
  its case in UserContextModal
