+++ packages/suite/src/components/suite/HomescreenGallery.tsx
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
 
 import { useDevice } from '@suite/device';
+import { notificationsActions } from '@suite-common/toast-notifications';
 import { Grid } from '@trezor/components';
 import {
     DeviceModelInternal,
@@ -12,10 +13,11 @@ import { borders, spacings } from '@trezor/theme';
 import { exhaustive } from '@trezor/type-utils';
 import { versionUtils } from '@trezor/utils';
 
-import { applySettings } from 'src/actions/settings/deviceSettingsActions';
+import { ConnectSubprocessModal } from 'src/components/suite/modals/ConnectSubprocessModal';
 import { getDefaultHomeScreenImage, getHomescreens } from 'src/constants/suite/homescreens';
 import { useDispatch } from 'src/hooks/suite';
 import { imagePathToHex } from 'src/utils/suite/homescreen';
+import { useConnect } from 'src/views/settings/SettingsDebug/useConnect';
 
 type HomescreensType = ReturnType<typeof getHomescreens>;
 type AnyImageName = HomescreensType[keyof HomescreensType][number];
@@ -44,6 +46,11 @@ const HomescreenImage = styled.img`
     border-radius: ${borders.radii.xs};
 `;
 
+// Lazy lambda — see ConnectWrapPlayground for context. Reading
+// `TrezorConnect.applySettings` at call time picks up the IPC-proxy override
+// installed by MainDesktop.tsx, instead of locking in the renderer stub.
+const connect = createConnect({ trezorConnect: TrezorConnect });
+
 type HomescreenGalleryProps = {
     onConfirm?: () => void;
 };
@@ -51,10 +58,11 @@ type HomescreenGalleryProps = {
 export const HomescreenGallery = ({ onConfirm }: HomescreenGalleryProps) => {
     const dispatch = useDispatch();
     const { device, isLocked } = useDevice();
+    const { start, subprocess } = useConnect(connect(TrezorConnect.applySettings));
 
     const deviceModelInternal = device?.features?.internal_model;
 
-    if (!deviceModelInternal) return null;
+    if (!deviceModelInternal || !device) return null;
 
     const isBitcoinOnlyFirmware = hasBitcoinOnlyFirmware(device);
     const setHomescreen = async (imagePath: string, image: AnyImageName) => {
@@ -90,22 +98,25 @@ export const HomescreenGallery = ({ onConfirm }: HomescreenGalleryProps) => {
     const isColorScreen = path.startsWith('COLOR');
 
     return (
-        <Grid gap={spacings.md} columns={4}>
-            {homescreens[deviceModelInternal].map(image => {
-                const src = resolveStaticPath(
-                    `images/homescreens/${path}/${image}.${isColorScreen ? 'jpg' : 'png'}`,
-                );
+        <>
+            <Grid gap={spacings.md} columns={4}>
+                {homescreens[deviceModelInternal].map(image => {
+                    const src = resolveStaticPath(
+                        `images/homescreens/${path}/${image}.${isColorScreen ? 'jpg' : 'png'}`,
+                    );
 
-                return (
-                    <HomescreenImage
-                        id={image}
-                        data-testid={`@modal/gallery/${path.toLowerCase()}/${image}`}
-                        key={image}
-                        onClick={() => setHomescreen(src, image)}
-                        src={src}
-                    />
-                );
-            })}
-        </Grid>
+                    return (
+                        <HomescreenImage
+                            id={image}
+                            data-testid={`@modal/gallery/${path.toLowerCase()}/${image}`}
+                            key={image}
+                            onClick={() => setHomescreen(src, image)}
+                            src={src}
+                        />
+                    );
+                })}
+            </Grid>
+            {subprocess && <ConnectSubprocessModal subprocess={subprocess} />}
+        </>
     );
 };
