+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/BackgroundGalleryModal.tsx
@@ -5,12 +5,13 @@ import { HomescreenGallery } from 'src/components/suite/HomescreenGallery';
 
 type BackgroundGalleryModalProps = {
     onCancel: () => void;
+    onConfirm?: () => void;
 };
 
-export const BackgroundGalleryModal = ({ onCancel }: BackgroundGalleryModalProps) => (
+export const BackgroundGalleryModal = ({ onCancel, onConfirm }: BackgroundGalleryModalProps) => (
     <Modal heading={<Translation id="TR_HOMESCREEN_GALLERY" />} onCancel={onCancel} width={600}>
         <Card>
-            <HomescreenGallery />
+            <HomescreenGallery onConfirm={onConfirm} />
         </Card>
     </Modal>
 );
